+++ android/src/main/kotlin/com/posthog/flutter/PosthogFlutterPlugin.kt
@@ -11,6 +11,7 @@ import android.util.Log
 import com.posthog.PersonProfiles
 import com.posthog.PostHog
 import com.posthog.PostHogConfig
+import com.posthog.PostHogOnFeatureFlags
 import com.posthog.android.PostHogAndroid
 import com.posthog.android.PostHogAndroidConfig
 import com.posthog.android.internal.getApplicationInfo
@@ -305,7 +306,14 @@ class PosthogFlutterPlugin :
 
                 sdkName = "posthog-flutter"
                 sdkVersion = postHogVersion
+
+                onFeatureFlags =
+                    PostHogOnFeatureFlags {
+                        Log.i("PostHogFlutter", "Android onFeatureFlags triggered. Notifying Dart.")
+                        invokeFlutterMethod("onFeatureFlagsCallback", emptyMap<String, Any?>())
+                    }
             }
+
         PostHogAndroid.setup(applicationContext, config)
     }