+++ lib/src/posthog_flutter_web_handler.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_dynamic_calls, avoid_annotating_with_dynamic
+
 import 'dart:js_interop';
 
 import 'package:flutter/services.dart';
@@ -31,6 +33,7 @@ extension PostHogExtension on PostHog {
   external void unregister(JSAny key);
   // ignore: non_constant_identifier_names
   external JSAny? get_session_id();
+  external void onFeatureFlags(JSFunction callback);
 }
 
 // Accessing PostHog from the window object