+++ test/posthog_test.dart
@@ -1,0 +1,35 @@
+import 'package:flutter_test/flutter_test.dart';
+import 'package:posthog_flutter/posthog_flutter.dart';
+import 'package:posthog_flutter/src/posthog_flutter_platform_interface.dart';
+
+import 'posthog_flutter_platform_interface_fake.dart';
+
+void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  group('Posthog', () {
+    late PosthogFlutterPlatformFake fakePlatformInterface;
+
+    setUp(() {
+      fakePlatformInterface = PosthogFlutterPlatformFake();
+      PosthogFlutterPlatformInterface.instance = fakePlatformInterface;
+    });
+
+    test(
+        'setup passes config and onFeatureFlags callback to platform interface',
+        () async {
+      void testCallback() {}
+
+      final config = PostHogConfig(
+        'test_api_key',
+        onFeatureFlags: testCallback,
+      );
+
+      await Posthog().setup(config);
+
+      expect(fakePlatformInterface.receivedConfig, equals(config));
+      expect(fakePlatformInterface.registeredOnFeatureFlagsCallback,
+          equals(testCallback));
+    });
+  });
+}